% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_add_quarto.R
\name{bhhi_add_quarto}
\alias{bhhi_add_quarto}
\title{Install BHHI Quarto Format}
\usage{
bhhi_add_quarto()
}
\description{
Installs the \href{https://github.com/ucsf-bhhi/bhhi-quarto}{BHHI Quarto format}
from GitHub.
}
