% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_cascade.R
\name{bhhi_cascade}
\alias{bhhi_cascade}
\title{Cascade Survey Objects with Full Contingencies}
\usage{
bhhi_cascade(.data, ..., .fill = "Overall")
}
\arguments{
\item{.data}{tbl A \code{tbl_svy} object}

\item{...}{Name-value pairs of summary functions}

\item{.fill}{Value to fill in for group summaries. Defaults to 'Overall'.}
}
\description{
Calculates full contingent statistics in the manner of \link[srvyr:cascade]{srvyr::cascade}.
}
\details{
When calculating a two-way contingency table (a.k.a. crosstab),
\link[srvyr:cascade]{srvyr::cascade} calculates proportions of the column variable for each level
of the row variable, but calculates the proportion of the row variable for
the overall sample. This makes two-way contingency tables difficult with
\link[srvyr:cascade]{srvyr::cascade}.

\code{bhhi_cascade()} instead calculates the proportion of the column variable for
the overall sample, which produces results that are compatible and comparable
to the results for the levels of the row variable
}
\examples{
data("nhanes", package = "survey")

survey_object = nhanes |>
  dplyr::rename(gender = RIAGENDR) |>
  dplyr::mutate(
    gender = factor(gender, 1:2, c("Male", "Female")),
    race = factor(race, 1:4, c("Hispanic", "White", "Black", "Other"))
  ) |>
  srvyr::as_survey(weights = WTMEC2YR)

survey_object |>
  srvyr::group_by(gender, race) |>
  bhhi_cascade(srvyr::survey_prop(proportion = TRUE))

survey_object |>
  srvyr::group_by(gender, race) |>
  srvyr::cascade(srvyr::survey_prop(proportion = TRUE), .fill = "Overall")

}
