% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_crosstab.R
\name{bhhi_crosstab}
\alias{bhhi_crosstab}
\title{Run a Weighted Crosstab}
\usage{
bhhi_crosstab(
  .data,
  row_var,
  col_var,
  pct_direction = "col",
  add_n = FALSE,
  vartype = c("se", "ci", "var", "cv"),
  level = 0.95,
  proportion = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{.data}{tbl A \code{tbl_svy} object}

\item{row_var}{Variable to show in the rows in output.}

\item{col_var}{Variable to show in the columns in output.}

\item{pct_direction}{\code{'col'} to calculate column percentages or \code{'row'} to
calculate row percentages. Defaults to \code{'col'}.}

\item{add_n}{Add cell N to output. Defaults to FALSE.}

\item{vartype}{Report variability as zero or more of: standard error ("se",
default), confidence interval ("ci"), variance ("var") or coefficient of
variation ("cv"). Defaults to hiding variability.}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{proportion}{Use methods to calculate the proportion that may have more accurate
confidence intervals near 0 and 1. Based on
\code{\link[survey]{svyciprop}}.}

\item{na.rm}{Drop missing values. Defaults to FALSE.}
}
\value{
A tibble with the crosstab results.
}
\description{
Simplifies running a weighted crosstab with \code{\link[=bhhi_cascade]{bhhi_cascade()}} and
\code{\link[srvyr:survey_mean]{srvyr::survey_prop()}}. Returns a tibble that can be directly customized or
formatted with \code{\link[=bhhi_reshape_crosstab]{bhhi_reshape_crosstab()}} and \code{\link[=bhhi_format_crosstab]{bhhi_format_crosstab()}}.
}
\examples{
data("nhanes", package = "survey")

survey_object <- nhanes |>
  dplyr::rename(gender = RIAGENDR) |>
  dplyr::mutate(
    gender = factor(gender, 1:2, c("Male", "Female")),
    race = factor(race, 1:4, c("Hispanic", "White", "Black", "Other"))
  ) |>
  srvyr::as_survey(weights = WTMEC2YR)

survey_object |>
  bhhi_crosstab(race, gender)

survey_object |>
  bhhi_crosstab(race, gender, pct_direction = "row", vartype = "ci")
}
