% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_format_crosstab.R
\name{bhhi_format_crosstab}
\alias{bhhi_format_crosstab}
\title{Create Formatted Table  From a Reshaped Crosstab}
\usage{
bhhi_format_crosstab(.data, decimals = 1)
}
\arguments{
\item{.data}{A tibble from \code{\link[=bhhi_reshape_crosstab]{bhhi_reshape_crosstab()}}}

\item{decimals}{An integer specifing the number of decimal places in the
results. Defaults to 1.}
}
\description{
Creates a formatted table via the \code{\link{gt}} package from a survey crosstab that has been reshaped via \code{\link[=bhhi_reshape_crosstab]{bhhi_reshape_crosstab()}}.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{data("nhanes", package = "survey")

survey_object <- nhanes |>
  dplyr::rename(gender = RIAGENDR) |>
  dplyr::mutate(
    gender = factor(gender, 1:2, c("Male", "Female")),
    race = factor(race, 1:4, c("Hispanic", "White", "Black", "Other"))
  ) |>
  srvyr::as_survey(weights = WTMEC2YR)

survey_object |>
  bhhi_crosstab(race, gender) |>
  bhhi_reshape_crosstab(race, gender) |>
  bhhi_format_crosstab()
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src='man_bhhi_format_crosstab_1.png' class='gt-example-img'>
}}
}

