% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_install_quarto_format.R
\name{bhhi_install_quarto_format}
\alias{bhhi_install_quarto_format}
\title{Install BHHI Quarto Format}
\usage{
bhhi_install_quarto_format()
}
\description{
Installs the \href{https://github.com/ucsf-bhhi/bhhi-quarto}{BHHI Quarto format}
from GitHub.
}
