% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_new_quarto.R
\name{bhhi_new_quarto}
\alias{bhhi_new_quarto}
\title{Create a New BHHI Quarto File}
\usage{
bhhi_new_quarto(filename = "bhhi-quarto.qmd", open = TRUE)
}
\arguments{
\item{filename}{Name of the new .qmd file.}

\item{open}{Open the new file in editor.}
}
\description{
Prepares a .qmd file with the BHHI template and ensures the \href{https://github.com/ucsf-bhhi/bhhi-quarto}{BHHI Quarto format} is installed.
}
