% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_read.R
\name{bhhi_rc_read}
\alias{bhhi_rc_read}
\title{Downloads REDCap Data}
\usage{
bhhi_rc_read(
  project_token_name,
  convert_factors = TRUE,
  for_stata = FALSE,
  col_types = "auto",
  export_survey_fields = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{project_token_name}{Name of the project's REDCap API Token as stored by
\code{\link[=bhhi_store_token]{bhhi_store_token()}} in the operating system credential vault or an
environment variable.}

\item{convert_factors}{Automatically convert categorical variables to factors
using the variable labels from the REDCap project. Default is TRUE.}

\item{for_stata}{Should the data be prepared for use in Stata? This makes
sure variable names are legal Stata variable names. Default is FALSE.}

\item{col_types}{Controls the variable types when parsing the REDCap data.
One of "auto", which uses the project metadata to assign variable types;
"guess", which uses the \code{\link[readr:parse_guess]{readr::col_guess()}} to determine the column
types; "string" which parses all variables as character strings; or a
custom \code{\link[readr:cols]{readr::cols()}} specification. Default is "auto".}

\item{export_survey_fields}{A boolean that specifies whether to export the
survey identifier field (e.g., 'redcap_survey_identifier') or survey
timestamp fields (e.g., instrument+'_timestamp'). The timestamp outputs
reflect the survey's completion time (according to the time and timezone of
the REDCap server.)}

\item{verbose}{Should messages be printed to the R console during the
operation. The verbose output might contain sensitive information (e.g.
PHI), so turn this off if the output might be visible somewhere public.}

\item{...}{Options passed to \code{\link[REDCapR:redcap_read_oneshot]{REDCapR::redcap_read_oneshot()}}.}
}
\value{
A tibble with the requested data.
}
\description{
Downloads data from REDCap by wrapping \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}}. It handles
variable type conversion and creating factors from categorical variables. It
can optionally rename variables to adhere to Stata's variable naming rules.
}
