% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_variable_labels.R
\name{bhhi_rc_variable_labels}
\alias{bhhi_rc_variable_labels}
\title{Parse REDCap Variable Labels}
\usage{
bhhi_rc_variable_labels(x)
}
\arguments{
\item{x}{Either a string with the name of the stored REDCap project API
token, or an \link[xml2:oldclass]{xml2 object} representing the
REDCap project XML.}
}
\value{
A tibble with three columns:
\itemize{
\item variable: the name of the variable as stored in REDCap
\item value: the integer value of the variable
\item label: the corresponding text label
}
}
\description{
Use the REDCap project metadata to parse the variable labels. Creates a
tibble with the values and labels for each categorical variable.
}
\details{
When provided with the name of an API token, it will download the project XML
from REDCap and then parse it.

When provided directly with the project XML, it will parse it.
}
