% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_read.R
\name{bhhi_redcap_read}
\alias{bhhi_redcap_read}
\title{Downloads REDCap Data}
\usage{
bhhi_redcap_read(
  project_token_name,
  for_stata = FALSE,
  col_types = NULL,
  export_survey_fields = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{project_token_name}{Name of the project's REDCap API Token. API tokens
must be stored in an environmental variable. Provide the name of that
environmental variable here.}

\item{for_stata}{Should the data be prepared for use in Stata? This makes
sure variable names are legal Stata variable names.}

\item{col_types}{Override default col_types for Stata by providing a
[readr::cols()] specification.}

\item{export_survey_fields}{A boolean that specifies whether to export the
survey identifier field (e.g., 'redcap_survey_identifier') or survey
timestamp fields (e.g., instrument+'_timestamp'). The timestamp outputs
reflect the survey's completion time (according to the time and timezone of
the REDCap server.)}

\item{verbose}{Should messages be printed to the R console during the
operation. The verbose output might contain sensitive information (e.g.
PHI), so turn this off if the output might be visible somewhere public.}

\item{...}{Options passed to [REDCapR::redcap_read()].}
}
\value{
A tibble with the requested data.
}
\description{
Downloads data from REDCap by wrapping [REDCapR::redcap_read()]. The API URL
is filled automatically, but can be overridden.
}
\seealso{
[redcap_token()]
}
