% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_do.R
\name{bhhi_stata_do}
\alias{bhhi_stata_do}
\title{Runs Stata REDCap data cleaning code.}
\usage{
bhhi_stata_do(data, do_file, quiet = TRUE, ...)
}
\arguments{
\item{data}{Tibble or data frame to pass to Stata.}

\item{do_file}{Path to the do file.}

\item{quiet}{Whether to suppress printing Stata output to the screen.}

\item{...}{Other options for controlling Stata interaction. Passed to
\code{\link[RStata:stata]{RStata::stata()}}}
}
\value{
A tibble with the cleaned data.
}
\description{
Runs the included code for cleaning the raw REDCap data download. A different
cleaning do file can also be used.
}
