% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_prep.R
\name{bhhi_stata_prep}
\alias{bhhi_stata_prep}
\title{Prepare Data for Stata}
\usage{
bhhi_stata_prep(data)
}
\arguments{
\item{data}{A tibble or data frame to prepare for Stata.}
}
\description{
Renames variables that don't comply with Stata's naming rules:
}
\details{
Stata names can only contain letters, digits, and underscores. Illegal
characters are replaced with underscores.

Stata names can't start with a digit. Names with leading digits are prepended
with an underscore.

Stata names can only be 32 characters long. Names are truncated to a maximum
of 32 characters.
}
