% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_store_token.R
\name{bhhi_store_token}
\alias{bhhi_store_token}
\title{Securely Store API Tokens}
\usage{
bhhi_store_token(token_name)
}
\arguments{
\item{token_name}{}
}
\description{
Uses the [keyring][keyring::key_set()] package to securely store API tokens
(e.g. REDCap) in your operating system's credential vault.
}
