% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_utils.R
\name{redcap_token}
\alias{redcap_token}
\title{Fetch appropriate REDCap API Token}
\usage{
redcap_token(project_token_name)
}
\arguments{
\item{project}{Either "main" (default) for the main REDCap project or "rds"
for the RDS project.}
}
\value{
The API token.
}
\description{
Fetches the project's REDCap API Token
}
\details{
To get your API token, ask one of the BHHI REDCap admins to grant you API
privileges and then request a token. See
https://redcap.ucsf.edu/api/help/?content=tokens for details.
}
