% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhhi_store_token.R
\name{bhhi_store_token}
\alias{bhhi_store_token}
\title{Securely Store API Tokens}
\usage{
bhhi_store_token(token_name)
}
\arguments{
\item{token_name}{Name of the token, a string}
}
\description{
Uses \link[keyring:key_get]{keyring} to securely store API tokens
(e.g. REDCap) in your operating system's credential vault.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bhhi_store_token("REDCAP_API_TOKEN")
\dontshow{\}) # examplesIf}
}
