% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_utils.R
\name{redcap_token}
\alias{redcap_token}
\title{Fetch appropriate REDCap API Token}
\usage{
redcap_token(project_token_name)
}
\arguments{
\item{project_token_name}{Name used to store the token via \code{\link[=bhhi_store_token]{bhhi_store_token()}} or the name of the environment variable with the token.}
}
\value{
The API token.
}
\description{
Fetches the project's REDCap API Token
}
\details{
To get your API token, ask one of the BHHI REDCap admins to grant you API
privileges and then request a token. See
https://redcap.ucsf.edu/api/help/?content=tokens for details.
}
\keyword{internal}
